Split core-plugins to their own directory
