+++ src/core-plugins/manage-metadata/manage-metadata.replacements
