+++ src/core-plugins/pillbox-menus/components/Menu.less
