+++ src/core-plugins/pillbox-menus/components/PillboxContainer.less
