+++ src/core-plugins/pillbox-menus/components/PillboxMenu.less
