+++ src/core-plugins/pillbox-menus/pillbox-menus.replacements
